# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AIQFome API",
        "description": "API para sistema de produtos favoritos",
        "contact": {
            "email": "admin@aiqfome.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Realizar login",
                "description": "Autentica um usuário e retorna um token de acesso",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "properties": {
                                                "key": {
                                                    "type": "string"
                                                },
                                                "text": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abc123..."
                                                },
                                                "role": {
                                                    "type": "string",
                                                    "example": "admin"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Realizar logout",
                "description": "Revoga o token de acesso do usuário autenticado",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou expirado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Obter perfil do usuário",
                "description": "Retorna os dados do usuário autenticado",
                "operationId": "e92c58f994e8be6e19813ce20425f83e",
                "responses": {
                    "200": {
                        "description": "Perfil recuperado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "properties": {
                                                "key": {
                                                    "type": "string"
                                                },
                                                "text": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Admin User"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "admin@example.com"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "admin"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou expirado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin-check": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verificar acesso de administrador",
                "description": "Verifica se o usuário autenticado tem permissão de administrador",
                "operationId": "fd0a452303896ad5476f626897bf66a9",
                "responses": {
                    "200": {
                        "description": "Acesso de administrador concedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "properties": {
                                                "key": {
                                                    "type": "string"
                                                },
                                                "text": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso negado - usuário não é administrador"
                    },
                    "401": {
                        "description": "Token inválido ou expirado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/token-status": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Verificar status do token",
                "description": "Retorna informações sobre o token de acesso atual",
                "operationId": "dc6d2fc3221a8ed865f49c90cb65c64c",
                "responses": {
                    "200": {
                        "description": "Status do token verificado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "properties": {
                                                "key": {
                                                    "type": "string"
                                                },
                                                "text": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "expires_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-11-15T23:00:00.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-16T23:00:00.000000Z"
                                                },
                                                "days_until_expiry": {
                                                    "type": "integer",
                                                    "example": 30
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou expirado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "message": {
                        "properties": {
                            "key": {
                                "type": "string"
                            },
                            "text": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "success",
                            "error"
                        ]
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "properties": {
                            "key": {
                                "type": "string"
                            },
                            "text": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "error"
                        ]
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}